requirements.txt
django
djangorestframework
mysqlclient
markdown
django-filter
bs4
codetiming
selenium
ipython
webdriver_manager
gekko
Pillow
django-filter
django-widget-tweaks
email-validator
aiohttp
django-admin-multiple-choice-list-filter
django-polymorphic
django-debug-toolbar
beautifulsoup4
PuLP
requests
